# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PUCPR Projeto Final - Webservices e MBaaS ",
        "description": "API criada como projeto final da matéria de Webservices e MBaaS - por Lucas Cândido dos Santos",
        "contact": {
            "name": "Lucas Cândido dos Santos"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/produtos/{id}": {
            "get": {
                "tags": [
                    "produtos"
                ],
                "summary": "Retorna um produto baseado no ID informado",
                "description": "Retorna um produto baseado no ID informado",
                "operationId": "04ff89b620ba602adef63569d15d944d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            },
            "put": {
                "tags": [
                    "produtos"
                ],
                "summary": "Atualiza um produto",
                "description": "Atualiza um produto",
                "operationId": "c45ef52a7b159513ec0c8878561a6c38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "example": "123"
                    },
                    {
                        "name": "produto",
                        "in": "query",
                        "description": "Código de referencia do produto",
                        "required": false
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome do produto",
                        "required": false
                    },
                    {
                        "name": "quantidade",
                        "in": "query",
                        "description": "Quantidade em estoque do produto",
                        "required": false
                    },
                    {
                        "name": "preco",
                        "in": "query",
                        "description": "Preço de venda do produto",
                        "required": false
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "description": "Tamanho do produto",
                        "required": false
                    },
                    {
                        "name": "cor",
                        "in": "query",
                        "description": "Cor do produto",
                        "required": false
                    },
                    {
                        "name": "inativo",
                        "in": "query",
                        "description": "Flag que define se o produto está ativo ou não",
                        "required": false
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "URL da imagem do produto",
                        "required": false
                    },
                    {
                        "name": "destaque",
                        "in": "query",
                        "description": "Flag que define se um produto é destaque ou não",
                        "required": false
                    },
                    {
                        "name": "preco_promocional",
                        "in": "query",
                        "description": "Preço promocional do produto",
                        "required": false
                    },
                    {
                        "name": "descricao",
                        "in": "query",
                        "description": "Descrição do produto",
                        "required": false
                    },
                    {
                        "name": "marca_id",
                        "in": "query",
                        "description": "ID da marca",
                        "required": false
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "description": "ID da categoria",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/produtos": {
            "post": {
                "tags": [
                    "produtos"
                ],
                "summary": "Cadastra um novo produto",
                "description": "Cadastra um novo produto",
                "operationId": "9c3a31cb5c64ba3c23f9ce329cea097c",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "query",
                        "description": "Código de referencia do produto",
                        "required": true
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome do produto",
                        "required": true
                    },
                    {
                        "name": "quantidade",
                        "in": "query",
                        "description": "Quantidade em estoque do produto",
                        "required": true
                    },
                    {
                        "name": "preco",
                        "in": "query",
                        "description": "Preço de venda do produto",
                        "required": true
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "description": "Tamanho do produto",
                        "required": true
                    },
                    {
                        "name": "cor",
                        "in": "query",
                        "description": "Cor do produto",
                        "required": true
                    },
                    {
                        "name": "inativo",
                        "in": "query",
                        "description": "Flag que define se o produto está ativo ou não",
                        "required": true
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "URL da imagem do produto",
                        "required": false
                    },
                    {
                        "name": "destaque",
                        "in": "query",
                        "description": "Flag que define se um produto é destaque ou não",
                        "required": true
                    },
                    {
                        "name": "preco_promocional",
                        "in": "query",
                        "description": "Preço promocional do produto",
                        "required": true
                    },
                    {
                        "name": "descricao",
                        "in": "query",
                        "description": "Descrição do produto",
                        "required": true
                    },
                    {
                        "name": "marca_id",
                        "in": "query",
                        "description": "ID da marca",
                        "required": true
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "description": "ID da categoria",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}